Moved CompressionUtils to shared module
