+++ collect_app/src/main/java/org/odk/collect/android/configure/qr/QRCodeScannerFragment.kt
@@ -12,8 +12,8 @@ import org.odk.collect.android.fragments.BarCodeScannerFragment
 import org.odk.collect.android.injection.DaggerUtils
 import org.odk.collect.android.projects.CurrentProjectProvider
 import org.odk.collect.android.storage.StoragePathProvider
-import org.odk.collect.android.utilities.CompressionUtils
 import org.odk.collect.androidshared.ui.ToastUtils.showLongToast
+import org.odk.collect.shared.CompressionUtils
 import org.odk.collect.settings.ODKAppSettingsImporter
 import java.io.File
 import java.io.IOException
